TJC + SimpleHAL v2.0 — แก้บั๊ก critical 24 รายการ + อัปเดตเอกสาร
## TJC Library
- เพิ่ม Page ID callback (TJC_PageIdCallback_t)
- DEFAULT pin config ปิด remap อย่างชัดเจน
- Static buffer เปลี่ยนเป็น local ป้องกัน re-entrant
- Error codes < 0x20 fallback ไป error callback
- TJC_FlushRxBuffer ใช้ RXNE interrupt disable ป้องกัน race
- SystemClockUpdate() เรียกอัตโนมัติใน TJC_Init
- Baud rate lock ไม่เกิน 115200

## SimpleHAL v2.0 — Critical Bug Fixes
- SimplePWM: TIM2 ใช้ RCC_APB1PeriphClockCmd (เดิม APB2 → PWM2 ไม่ทำงาน)
- SimplePWM: ป้องกัน NULL dereference + divide-by-zero + multi-channel timer base
- SimpleUSART/I2C/SPI: เพิ่ม RCC_APB2Periph_AFIO ก่อน GPIO_PinRemapConfig
- SimpleSPI: แก้ remap pin (PC6/PC8/PC7/PC5 → PD1/PD2/PD3/PD0)
- SimpleSPI: สลับค่า speed enum ให้ตรงความเร็วจริง
- SimpleDelay: ลบ constructor (ทำงานก่อน SystemCoreClockUpdate) + runtime guard
- SimpleADC: เปิด TSVREFE + auto-init ตั้ง GPIO + divide-by-zero protection
- SimpleFlash: WriteWordWithErase 1 erase (เดิม 4)
- SimpleFlash: latency ปรับอัตโนมัติตาม SystemCoreClock
- SimplePWR: ลบ ARM-only NVIC->SCTLR (ใช้ไม่ได้บน RISC-V) + PA0 WKUP note
- SimpleTIM: divide-by-zero protection
- SimpleTIM_Ext: Time_t uint8_t→uint16_t (RAW mode >255) + Stopwatch guard
- SimpleDMA: busy-wait timeout check + volatile fix
- SimpleI2C: NACK detection (I2C_FLAG_AF) — scan เร็ว ~10 เท่า
- SimpleGPIO: เพิ่ม PD0(18), PD1(19) ใน pin enum
- SimpleOPAMP/SimpleGPIO: volatile fixes
- SimpleUSART: INT32_MIN overflow fix
- SimpleArduino: PrintFloat buffer overflow fix
- Simple1Wire: IRQ depth counter ป้องกัน nested disable
- SimpleWWDG: PCLK1→SystemCoreClock runtime (เลิก hardcode 24MHz)
- SimpleIWDG: reload clamp ≥ 1
- SimpleADC.h: เปลี่ยน alias enum ซ้ำ → #define

## เอกสาร
- อัปเดต readme/ ทั้ง 16 ไฟล์ + AGENT.MD
- เพิ่ม SimpleHAL v2.0 breaking changes table
- เพิ่มตาราง package compatibility (TSSOP-20/SOP-8/SOP-16)

+++ User/SimpleHAL/SimpleTIM_Ext.h
@@ -83,8 +83,8 @@ typedef enum {
  */
 typedef struct {
     uint16_t hours;      /**< ชั่วโมง (0-1193 สำหรับ max time) */
-    uint8_t minutes;     /**< นาที (0-59 normalized, 0-255 raw) */
-    uint8_t seconds;     /**< วินาที (0-59 normalized, 0-255 raw) */
+    uint16_t minutes;    /**< นาที (0-59 normalized, 0-71582788 raw) */
+    uint16_t seconds;    /**< วินาที (0-59 normalized, 0-4294967295 raw) */
 } Time_t;
 
 /* ========== Buffer Size Recommendations ========== */
